requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
annotated-types==0.7.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
bounded-pool-executor==0.0.3
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
docstring_parser==0.17.0
einops==0.8.1
exceptiongroup==1.3.0
executing==2.2.0
filelock==3.13.1
fonttools==4.58.2
frozenlist==1.7.0
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.44
h5py==3.14.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.4
jsonargparse==4.40.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
lazy_loader==0.4
lightning==2.5.2
lightning-utilities==0.14.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.4
nest-asyncio==1.6.0
networkx==3.3
numpy==2.1.2
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
orjson==3.10.18
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.8
pqdm==0.2.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.2
PyYAML==6.0.2
pyzmq==26.4.0
requests==2.32.4
rich==14.0.0
scikit-image==0.25.2
scipy==1.15.3
sentry-sdk==2.29.1
setproctitle==1.3.6
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
sympy==1.13.3
tensordict==0.9.0
tifffile==2025.5.10
torch==2.7.1+cu118
torch-kmeans==0.2.0
torchaudio==2.7.1+cu118
torchinfo==1.8.0
torchmetrics==1.7.2
torchvision==0.22.1+cu118
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
triton==3.3.1
typeshed_client==2.8.2
typing-inspection==0.4.1
typing_extensions==4.12.2
urllib3==2.4.0
wandb==0.20.1
wandb_osh==1.2.2
wcwidth==0.2.13
yarl==1.20.1
zipp==3.23.0
